build(hardening): add hardening_level option with compiler warning-as-error flags
Add meson option 'hardening_level' (none/basic/strong/maximum) controlling
compiler hardening flags:

- basic (default): stack protector + FORTIFY_SOURCE + libc++ hardening
- strong: + -Werror=return-type, -Werror=uninitialized,
  -Werror=format-security, -Werror=implicit-function-declaration,
  -Wformat=2 (warning, not yet error due to ~65 existing mismatches)
- maximum: + -ftrapv, -Wformat=2, _GLIBCXX_ASSERTIONS on Linux

Warning flags use add_project_arguments to avoid breaking subprojects
like anet.

Fixes caught by the new flags:
- anet: format mismatches (clock_t/size_t as %%d), uninitialized va_list
- dbgen: size_t as %%d, int as %%lx, time_t as %%lld
- ColorSet: uint32 as %%ld
- gfx/spritesys: uint16 as %%ld, size_t as %%d
- SpriteGroupList, tileset, workmap, controlpanelwindow,
  greatlibrarywindow: sprintf with non-literal format strings
- c3endgamedb, C3UnitDB: uninitialized pointer returns
- C3World: missing return values in non-void functions

+++ ctp2_code/gs/dbgen/ctpdb.cpp
@@ -34,12 +34,12 @@
 // - Variable 'or' renamed, because this a reserved symbol (same as ||) now.
 // - Modifed db_add_bit_pair function to allow bit pairs to have default
 //   values so that when two records are merged, only the bit is merged
-//   in that is set. - Sep. 28th 2004 Martin Gühmann
+//   in that is set. - Sep. 28th 2004 Martin Gï¿½hmann
 // - Portability improved (ctplinuxfan).
 // - Added method to determine the number of entries that should be parsed
-//   in order to support the old pollution database. (July 15th 2006 Martin Gühmann)
-// - Added map.txt support. (27-Mar-2007 Martin Gühmann)
-// - Added Const.txt support. (29-Jul-2007 Martin Gühmann)
+//   in order to support the old pollution database. (July 15th 2006 Martin Gï¿½hmann)
+// - Added map.txt support. (27-Mar-2007 Martin Gï¿½hmann)
+// - Added Const.txt support. (29-Jul-2007 Martin Gï¿½hmann)
 // - Reported failure to create a (writable) destination file in copy_file
 //
 //----------------------------------------------------------------------------
@@ -366,7 +366,7 @@ void db_end_record(char *name)
 	FILE * stamp = fopen(filename, "w");
 	Assert(stamp);
 	if(stamp) {
-		fprintf(stamp, "//%" PRId64 "\n", time(0));
+		fprintf(stamp, "//%ld\n", (long)time(0));
 		fclose(stamp);
 	}
