+++ ctp2_code/libs/anet/demo/utils/logpkt.c
@@ -153,7 +153,7 @@ void logPkt(FILE *outFile, const void *buffer, size_t length, unsigned long peer
 	}
 	*p++ = 0;
 
-	fprintf(outFile, "%2s %10d %2d %3d : %s\n", tag, eclock(), peer, length, tbuf);
+	fprintf(outFile, "%2s %10lu %2lu %3zu : %s\n", tag, (unsigned long)eclock(), peer, length, tbuf);
 }
 
 #endif
